+++ MalkovaMM/lab7/lab7/pol.h
@@ -1,12 +1,14 @@
 #ifndef STRUCU_H
 #define STRUCU_H
+#define MAX_LEN 1000
 
 typedef struct {
 	int* coef;
 	int deg;
 }polinom;
 
-polinom pcreate(int deg);
+int degpol(const char* fname);
+polinom pcreate(const char* fname, int deg);
 polinom pplus(polinom* p1, polinom* p2);
 polinom pminus(polinom* p1, polinom* p2);
 polinom pumn(polinom* p1, polinom* p2);
@@ -14,3 +16,4 @@ int pznach(polinom* p, double  x);
 polinom pdif(polinom* p);
 
 #endif
+
