+++ MalkovaMM/lab7/lab7/task7.txt
@@ -1,0 +1,2 @@
+3;6;9;1;0;6
+7;1;12;2;1
