--- MalkovaMM/labb7/labb7/pol.h
@@ -1,18 +1,0 @@
-#ifndef STRUCU_H
-#define STRUCU_H
-#define MAX_LEN 1000
-
-typedef struct {
-	int* coef;
-	int deg;
-}polinom;
-
-int degpol(const char* fname);
-polinom pcreate(int deg);
-polinom pplus(polinom* p1, polinom* p2);
-polinom pminus(polinom* p1, polinom* p2);
-polinom pumn(polinom* p1, polinom* p2);
-int pznach(polinom* p, double  x);
-polinom pdif(polinom* p);
-
-#endif
