--- MalkovaMM/labb7/labb7/task7.txt
@@ -1,1 +1,0 @@
-3;6;9;1;0;6
